requirements.txt
streamlit
requests
pypdf
python-docx
beautifulsoup4